add kci_admin.lab and move all lab-related code there
Keep the kci_add_lab script to the bare minimum and make all the code
available in the kci_admin package by moving all the lab-related code
to a new module kci_admin.lab.

Keep a generic kci_admin.request() function to create the request data
for a given API host using the settings.  Which can be reused by any
commands interacting with the API.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>